Fix a small glitch to correct codegen result
Only tuple binding and construction is valid now

One example to try in shell:

$ stack run repl
$ > :gen let (?a, ?b, _) = (20, 3, "helloworld") in a

and paste output in a text file, invoking clang on it.

$ echo $?

you will get 20. This example shows how the whole machine
gets running.

An extra bonus is we don't need to write all codegen
at once but to do it progressively.

For completed part, it will use type system, builtin of LLVM, to
do a basic type checking. For unfinished part, it will
throw undefined error and exit repl and this is a intended.

Contribution is welcome!!